refactor(notify)!: rename receivers to recipients
As I learned over time as a non-native English speaker, the word "receiver" is
not used in the context of "recipient" in English. I think it's better to use
"recipient" instead of "receiver" in the context of notification receivers.

To take the next steps with this project, I think it's better we introduce
this change now, and form a even more solid foundation for the future.